Update carrier move event

+++ aggregate/carrier/state.go
@@ -7,8 +7,8 @@ import (
 )
 
 type PositionState struct {
-	X int
-	Y int
+	X float64
+	Y float64
 }
 
 type status int
